[DM-22013] Rework of CADC TAP chart
Increment version.

Rename db to mock-qserv.  Add option in values.yaml to allow option
for using mock-qserv or not.  Rename schema-db to tap-schema-db.

Get consistent about using labels in the chart using the labels
helper everywhere.

Use the new image names in the DM-22013 branch of lsst-tap-service,
and allow the docker tag to come from values.yaml for all images as
a consistent set.