[DM-22013] Pass in secrets from command line
This is about how to inject the google_creds.json file which is a
secret in using helm.  First, add an empty string for the placeholder
secret in values.yaml.  If no google_creds override is passed in,
an empty secret is put in (which means you can't use async).  Otherwise,
you can have a values.yaml that has the secret injected as a base64
string, or you can use a helm v2.10+ version's --set-file option to
inject the secret from a file.

So for this, it would be:

helm install lsstsqre/cadc-tap --set-file google_creds=google_creds.json

This would make the secret with the file you pass in on the command line.