[DM-22297] Update nublado chart for -int
Okay in this round I fix some issues found when deploying these
changes on google, with the deployment spec selector complaining,
because it was on the wrong indent level for some reason.

Add the ability to have the ingress be more configurable from values
yaml file, so that -int and -stable can override this for auth.