[DM-22297] Nublado static files via values.yaml
Instead of using static files, which were always kind of a pain,
I've figured out how to just have blocks of JSON or YAML inside
of the values.yaml file, which can be easily overridden.  So
just pass them in.  Also, apparently the files on disk were base64
encoded, so that was really wrong anyway.