[DM-22532] Nublado: Fix home pv creation
There was a bug where you'd get error messages about problems creating
the home physical volume due to annotations being wrong.  So that must
be some tighter checking done by helm or kubernetes now, so that was
in the wrong place.  So moved that into the metadata section, and now
I get an error when doing a kubectl create (but not apply, which is
interesting in its own right) if the GID isn't quoted.  So I'm going
to quote that just to be careful.