[DM-23217] Copy over opendistro-es chart for serving
Okay go from the LSP way to using the opendistro-es chart.  I'm copying
in version 1.1, and it comes from here:

https://github.com/opendistro-for-elasticsearch/community/tree/master/open-distro-elasticsearch-kubernetes/helm

Copied at hash:

710e8e022595a8f3c0a7897eadf46074d0a03808

This is because they don't seem to publish this chart over the normal Helm HTTP way,
so I'm going to add it to our repository and serve it out of there for now.