[DM-23877] Update Gafaelfawr chart 1.0.0 release
Rename secret paths, the default Docker image, and the path to the
configmap.  Update URLs.  Change the names of some additional
configuration settings.  Simplify the structure of values.yaml.
Update the URL for checking authorization for the /auth/tokens
routes.

Add the /logout route and the after_logout_url configuration
parameter.  Reorganize the ConfigMap to match the upstream
configuration file ordering.  Rename the configuration file it
creates to match the new upstream naming.

Default to the new /login route that we want to use, but allow
environments with the older /oauth2/callback route configured
to override this.