[DM-23047] Explicitly set java max heap size
So while we have limits for the pod, java also has its own set of
limits, which is annoying.  So there's the max heap size for java,
which has some black magic for determining what it should be, such
as the available memory on the machine.

I would have liked to have used the resource limit, but since that
may not exist, you can't easily index into it even when providing
a default if some level of the key doesn't exist.  So that's why
I've put it as a separate variable.  Also this should probably be
slightly less than the limits of the pod, since java itself and
other container processes will have some overhead.  I've set the
max to 4gb, which is the max the heap should grow to.  If this is
too high, we can set this lower, and override it in stable.