Add Gafaelfawr support for password-protected Redis
Since we're giving users a shell into the Kubernetes private
network, add password protection to Gafaelfawr's Redis for an
additional layer of protection.

Change the pull policy for the Gafaelfawr image to IfNotPresent
since we're pulling tagged images and the tag won't move.  Users
that override this to point to a ticket branch can also override
the pull policy.