Add a Redis liveness probe
Every thirty seconds, make sure that Redis can write to its
database.  This will catch problems we've seen in the past where
the NFS mount for the persistent volume goes read-only due to NCSA
disk issues.  If that happens with this probe, the container will
be killed and restarted, which is the correct action to take to
recover.

Set the restart policy to OnFailure to allow for this restart.