Add TLS configuration to nubaldo ingress
For the base and summit clusters, we're not going to run the ingress
or the cert-manager installation.  Those instead will be handled by
IT.  However, the default cluster issuer will be configured for
lsst.org, rather than for our lsst.codes hostnames.

Add support to the nublado chart for configuring TLS for the hostname
used by the JuptyerHub proxy ingress.  ingress-nginx should then
merge that TLS configuration with all other ingress configurations
for the same hostname that don't specify a TLS configuration, which
is all of them at present.  That should accomplish the goal of using a
separate certificate issuer for the nublado hostname.  That issuer
will be set up by lsp-deploy.

Take the issuer name and the secret name as configuration settings in
values.yaml.