Don't reuse proxy.ingress.host for TLS
We have a weird hybrid configuration where some of our environments
use named virtual hosts and some of them use the wildcard host.
Using proxy.ingress.host requires switching everything away from the
wildcard, and while that's probably the right thing to do, it's
unrelated to this change.  Add a separate tls_host configuration to
specify which hostname TLS certificates will be requested for, so that
we can leave proxy.ingress.host unset and continue to use the wildcard
host.