[DM-27245] Nublado2 clusterrole
This gives the hub role a lot more things it can do.  We need to be
able to use a clusterrole to create namespaces and watch pods in
other namespaces, so we'll need this for the multinamespace work.

This is also required to create any kubernetes objects from the
hub, so the arbitrary resource creation objects have to be of
types that can be created by the hub, otherwise you get a 403.

The list of roles comes from what is currently on nublado, so we
can trim it down, or keep it here and trim it once we get other
parts working.

Note that the clusterrolebinding binds to the service account that
the subchart creates.