Allow nginx-ingress to talk to the proxy
We use a custom ingress running as a pod in the ngninx-ingress
namespace, so we need special NetworkPolicy configuration to allow
it to talk to the proxy pod.  This will be a little simpler when
we're running the current version of the Zero to JupyterHub chart,
which has built-in support for this.  In the current version, it
requires a custom rule.