Fix Gafaelfawr auth-url annotation for tokens
Using an unqualified gafaelfawr-service.gafaelfawr doesn't work
because nginx uses a Lua DNS library that doesn't honor the search
option in /etc/resolv.conf.

We had previously not used svc.cluster.local because of some issue
on the summit, but inspecting the summit DNS configuration now, it
appears to use cluster.local as its default domain and thus
svc.cluster.local should work.  Return to that configuration,
which will hopefully work on all of our environments.

This problem is temporary since Gafaelfawr 2.0.0 will remove the
need for this annotation in the Gafaelfawr chart (although we'll
still need something similar in the charts of other protected
services).