[DM-28482] Nublado2: default empty image list
Add in an empty list for the other images to show.  Otherwise
nublado2 will get unhappy when showing the options form and trying
to find the default list of images to show.  By default, there is
nothing in there.  But someone could use this list instead of
cachemachine, or merge it with those results.