Update Nublado 2 auth configuration
We can now use the same auth-url on all environments, so set it to
the gafaelfawr-service.gafaelfawr.svc.cluster.local URL so that
instantiations don't have to set it.  Pass only the token into
Nublado 2, not other headers, since we'll use an API call to retrieve
the other data.  Add a base_url config setting that has to be
set on all instantiations, which will be used to construct the URL
to contact Gafaelfawr and retrieve token metadata.