Security hardening for cachemachine chart
Put the external ingress for cachemachine behind Gafaelfawr to
prevent the Internet from talking to it without authentication.
We will eventually also want to restrict access to it from inside
the cluster, but that's much less important and can be tackled via
a NetworkPolicy later.

Add pod and container hardening best practices from SQR-048, which
should be a no-op for this service since it doesn't do anything
exciting other than answer HTTP requests and make Kubernetes API
calls.