[DM-28482] pod_cmd instead of cmd
The zero-to-jupyterhub charts normally set the command to
jupyterlab-singleuser, and override what the command is for
the docker container.  If you set cmd = , this means use
the default command for the docker container entrypoint.
This will allow the chart to configure the container command line,
if needed.  Defaulting to the container default.

Unforunately, trying to unset something in a subchart doesn't
seem to work well, or at least, I can't get it to work.  So
I'm adding a new way to set this, which has a default of none,
called pod_cmd.