Refresh Gafaelfawr charts with best practices
Update the template helpers to use the chart name, since apparently
the definitions are global within the chart and its subcharts, and
update the references accordingly.

Use standard labels on all resources.

Use standard selector labels plus a component label as the selectors
for the deployments instead of an app label.

Quote more strings in the templates.