Convert Gafaelfawr Redis to a StatefulSet
This is apparently the correct way of handling dynamic provisioning
of persistent volume claims, since it prevents the deadlock problems
when restarting a deployment with ReadWriteOnce persistent volumes.

Default to dynamic provisioning of a 1GiB volume (current NCSA
stable uses less than 10MiB).  Continue to support both an existing
PVC name (used at NCSA) and emptyDir (suitable for minikube and other
test deployments).

Bump the chart version to 2.0.0, since this is a major, breaking
change.  Give up on trying to match the chart version with the
application version, since the Helm chart and Gafaelfawr don't have
breaking changes at the same time or frequency.