[DM-29532] Update chart for refactors
We're folding in all the options form parts into the lower level,
and adding a leading 0 because there's trickiness I think with the
parsing and figuring out if it's a string or a float if it's a leading
zero.  Or at least I think I've had problems like this before.