[DM-29647] Refactor automatically generated resources
We're moving the resources we need and create, like the service
account and lab environment, into the chart.  Helm doesn't like
to merge and extend lists though.  So we put the ones the helm
chart generates here, and we can extend it in the nublado config
part for additional volumes per environment.