Support new Gafaelfawr secret in mobu
Provide a new configuration parameter, gafaelfawrSecret, that
specifies the name of a secret in the mobu namespace that will
contain the Gafaelfawr token.  If this is set, do not create the
Gafaelfawr vault secret in the mobu namespace or mount it as a
volume.

Once all installations have switched to Gafaelfawr 2.0, these
portions of the template can be deleted.

Also require Helm v3 for the chart (we're using that everywhere
already) and bump the mobu appVersion in anticipation of the mobu
release with GAFAELFAWR_TOKEN support.