Update nublado2 for Gafaelfawr 2.0
Mount the new Gafaelfawr token secret into the hub container as
well.  This means this version of the nublado2 chart can only be
deployed in environments that Gafaelfawr 2.0 has been deployed in.
Unfortunately, because most of the work in the chart is done in
values.yaml, we can't do better than that.

Change the scope used to protect moneypenny.

Change the nublado2 and moneypenny charts to require Helm 3.  We're
already using that everywhere and know that it works with the
upstream chart, even though the upstream chart is still v1.