Add post-install hook for database initialization
Rather than having Gafaelfawr try to initialize the database every
time the container starts, run a Helm post-install (but not
post-upgrade) hook to do this.  Rename the Kubernetes secrets
post-install hook so that it doesn't conflict, and update the
network policy.

Add comments to the network policy to explain what it is doing.