Add support for stopping Gafaelfawr sidecars
The Cloud SQL Auth Proxy runs forever, which causes jobs (including
in cron jobs) to apparently never finish.  Adopt a technique from
https://stackoverflow.com/questions/41679364/kubernetes-stop-cloudsql-proxy-sidecar-container-in-multi-container-pod-job
to run the Cloud SQL Auth Proxy in a monitor shell script and signal
it via a shared emptyDir volume when the main portion of the job is
complete.