Support a NetworkPolicy for Gafaelfawr
If enabled, this will restrict access to the Gafaelfawr pod to only
pods in a namespace tagged with gafaelfawr.lsst.io/ingress with a
value of "true".  This will require all cluster access go through
the ingress and will allow hiding of routes by not creating an ingress
for them (the plan is to use this for /auth to prevent users from
minting arbitrary internal and notebook tickets).