Update firefly chart to SUIT 2.2.0
Change appVersion to 2.2.0 and set the default image tag to the
appVersion rather than hard-coding it in values.yaml.  Switch the
chart to Helm v3 (which we're already using to install it in
Science Platform environments).

This chart should be renamed to suit or portal, but that will be
done later as part of a larger overhaul.