Support Gafaelfawr custom resource
Add a custom resource definition for GafaelfawrServiceTokens,
implemented via the new controller in Gafaelfawr 3.0.0.  Deploy that
controller instead of the cron job and post-install hook that
created secrets based on the Gafaelfawr configuration, and drop
support for the now-obsolete Gafaelfawr configuration block to
specify the secrets to create.

Convert mobu and nublado2 to use GafaelfawrServiceToken custom
resources for their service token secrets, and delete the remnants
of the pre-Gafaelfawr-2.0.0 support for mounting Gafaelfawr's
signing key to forge JWTs.

Fix the securityContext for the Gafaelfawr CloudSQL sidecar
container.  The securityContext stanza was present twice, which
meant only the second one had any effect.