Mount volume for .next/cache/images
During runtime, the next/image component creates cached versions of
image variants that the app serves. These are stored
in /.next/cache/images. Since we run squareone in a read-only filesystem
for security, the next app isn't able to create and write to this
directory. A solution is to mount an emptyDir at that path. This change
involves an update to the squareone Helm chart.