Let mobu talk to cachemachine
Bump version of mobu so that it uses its own token and not the user
token for talking to cachemachine, and give it exec:admin scope so
that it will have access.  We can tighten this later, but mobu is
fairly safe to give enhanced permissions to (and admin:token is
already stronger than exec:admin).