Add vo-cutouts chart
Add a new chart for the vo-cutouts service and its supporting Redis,
database worker, and cutout worker.  The cutout worker pod is based
on the stack container with Dramatiq and the cutout service deployed
dynamically on container start.

Add read:image to the Gafaelfawr list of known scopes, since it will
be used for access control for the vo-cutouts service.