alert-stream-simulator: create static topic via job, not resource
The Kafka topic that holds static alert data needs to be created de novo and
populated with a single pass of the sample alert data.

Populating it with data is done via a Kubernetes Job. This job runs on *every*
deployment of the simulator, which ensures that we're always using the latest
schema and alert data.

But it also means we need to delete the topic before filling it with data.
There is no clear way to signal this to Strimzi, so making it a Strimzi
KafkaTopic resource doesn't really work; instead, this revision uses a new
version of the stream simulator with configuration set to always create a new
topic, overwriting one if it already exists.