Rewrite Moneypenny Helm chart
Use the current template and layout and add standard labels to
all of the resources.

Add a NetworkPolicy that prevents anything other than the ingress
from talking to the web service so that we can enforce
authentication.

Move some configuration, such as the ingress annotations, into
the ingress resource since we don't intend to support customization.

Update the ingress to the current API version.

Remove the VaultSecret resource in favor of having Moneypenny always
use its configured imagePullSecrets.