Update all Ingresses, fix NetworkPolicies
Update all Ingress objects to networking.k8s.io/v1.  This is the
only supported API in Kubernetes 1.22 and later, and in ingress-nginx
4.0 and later.  We are dropping support for Kubernetes versions
prior to 1.18, when this API was added.

Fix the NetworkPolicy created by the rsp-starter to restrict access
except from the NGINX ingress as intended, and fix all the
NetworkPolicy objects created from that template.  The prior version
was intended for use with a Redis rather than a web application.