Rework cadc-tap chart to use current conventions
The instigating bug was that the UK folks wanted to be able to
specify a custom Docker image, but the Docker image was hardcoded
in the chart.  Fix that by reworking the cadc-tap chart to use our
current chart conventions and values.yaml layout.

This drops support for using direct Kubernetes secrets in favor
of assuming Vault secrets.