Update moneypenny chart
Update for the 1.0.0 release.  Add permissions to list and watch
pods so that we can use a watch to wait for the provisioning pod
to finish.  Increase the NGINX proxy timeout to 5m plus ten seconds,
which is long enough to wait for the default timeout for a pod to
run.