Update cachemachine Ingress configuration
Separate the cachemachine Ingress into an authenticated one and an
unauthenticated one, in preparation for making the Nublado query to
cachemachine go through the ingress so that we can add
NetworkPolicy protection.  The net effect of this will prevent anyone
inside the cluster from telling cachemachine to cache arbitrary
images, which could pose a security concern.