Make NetworkPolicy unconditional
Remove the boolean setting controlling installation of the
NetworkPolicy for gafaelfawr, moneypenny, and vo-cutouts and
instead always install it.  This requires that the ingress
properly be tagged with the Gafaelfawr label.  Our new starter
makes the NetworkPolicy installation automatic if the Ingress
is enabled, so may as well make that match across all of our
charts.