Fix security settings for cadc-tap databases
Run mock-qserv as root with a writable file system for now, since
the startup of its Docker image apparently requires rewriting files
in /etc (why?).  Fixing this looks like a larger issue.

Fix the user the uws-db container runs as to hopefully match the
UID and GID of the postgres user and thus solve initdb issues.