Limit image cutout workers to one process/thread
I believe that Butler is not thread-safe, so limit each worker to
a single thread since it uses a global Butler instance so that the
SQL connection is cached.  Limit the worker to a single process,
since we want to scale horizontally by adding more pods rather than
running multiple processes in a single pod.