Fix vo-cutouts secret permissions without privs
The fix-secret-permissions container doesn't need to run as root
and can still drop permissions because the problem that it's
working around is that the secret files are world-readable.  It
can therefore run as the same user as the regular container to
copy the secret files over and make them readable only by that
user.