[DM-34433] Sherlock: Set cors headers
Okay so this should inject the cors headers, and I'll get rid of the
code that generates the cors headers from the sherlock code.

Unfortunately, setting the list of domains doesn't seem to work.
This means that I can't set the host to just status.lsst.codes, instead
the default is *, which isn't great, but isn't terrible.

nginx.ingress.kubernetes.io/cors-allow-origin is the annotation
that is supposed to work, but when I set it it does not work.