Change default branch to main, update docs
Change all references from master to main, and enable testing on
pull requests to main rather than pushes to any branch except
master and gh-pages.

Delete a bunch of obsolete details from the documentation that
referred to the really old way we managed charts.

Update the version of Python used for testing to 3.11.